change Major Tick  Step Hint to Integer property. This may cause warning in reading old OPI files which have widgets with this property.

+++ applications/plugins/org.csstudio.opibuilder.widgets/src/org/csstudio/opibuilder/widgets/editparts/AbstractScaledWidgetEditPart.java
@@ -94,7 +94,7 @@ public abstract class AbstractScaledWidgetEditPart extends AbstractPVWidgetEditP
 					final Object newValue,
 					final IFigure refreshableFigure) {
 				AbstractScaledWidgetFigure figure = (AbstractScaledWidgetFigure) refreshableFigure;
-				figure.setMajorTickMarkStepHint((Double) newValue);
+				figure.setMajorTickMarkStepHint((Integer) newValue);
 				return true;
 			}
 		};
